Fixed Debian release logic & Fixed dependcies
Using https://enterprise.proxmox.com/debian/ for gpg downloads if needed, as trixie is on there, not the https://downloads.proxmox.com/debian/

Using bookworm as the default for debian, as trixie isn't final yet.

using apt-get as that is what is suposed to be used in scripts

removed the hard coded debian in DockerFile to use selected debian release.

changed the version patch from pve-relaxablermrr to -pve-kernel-builder to show that the kernel is from -pve-kernel-builder. Especially since relaxablermrr is built into newer kernels.

Fix dependancies for bookworm and trixie